docs: add design doc

+++ src/parser.rs
@@ -9,7 +9,7 @@ use nom::{
 
 #[derive(Clone, Debug, PartialEq)]
 pub enum Value<'a> {
-    Number(f64),
+    Number(u64),
     Identifier(&'a str),
     String(String),
     SExpr(Vec<Value<'a>>),
